bt-joiner: only enable the hci interface specified by the config
Drop the auto-enable policy as that forces bluez to automatically
initialize and use all the interfaces available in the system.
Add back the hciconfig up/reset commands to only enable and use the
interface specified by the config.

hciconfig (deprecated) will be replaced by btmgmt once that is packaged
and available in Alpine.

Signed-off-by: Ricardo Salveti <ricardo@opensourcefoundries.com>