bluetooth_6lowpand.sh: remove set -e
The current code is already handling the return code for every command
it uses, so no need to use 'set -e' anymore (otherwise the entire error
handling logic would need to be changed, as it is not simple to check
for return codes when '-e' is used).

This script will later be moved to Python once the new BT 6LoWPAN kernel
API gets merged.

Signed-off-by: Ricardo Salveti <ricardo@opensourcefoundries.com>