radvd64: remove embedded interface-monitor
Right now we bundle the interface-monitor into the radvd container
so that BLE 6lowpan can use it to dynamically add the bt0 IP address
whenever the bt0 interface is around.

However, we also need to add an IP address to the OpenThread interface
so that services like DNS and CoAP proxy have static settings on the
IoT nodes.

Let's remove interface-monitor from radvd64 so that it can be split
out into it's own container. (OpenThread doesn't use radvd).

Signed-off-by: Michael Scott <mike@foundries.io>